refactor: Ladder 클래스 수정 및 패키지 구조 변경 - Ladder 클래스가 PlayersAndWinningContents 포함하고 있도록 수정 - 패키지 구조 수정
