+++ src/main/java/laddergame/domain/ladder/Ladder.java
@@ -1,7 +1,6 @@
 package laddergame.domain.ladder;
 
 import laddergame.domain.HeightOfLadder;
-import laddergame.domain.PlayersAndWinningContents;
 import laddergame.domain.ladder.strategy.LinkStrategy;
 
 import java.util.Collections;
@@ -10,9 +9,11 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class Ladder {
+    private final PlayersAndWinningContents playersAndWinningContents;
     private final List<Line> lines;
 
     private Ladder(HeightOfLadder heightOfLadder, PlayersAndWinningContents playersAndWinningContents, LinkStrategy linkStrategy) {
+        this.playersAndWinningContents = playersAndWinningContents;
         this.lines = Stream.generate(() -> Line.newLine(playersAndWinningContents, linkStrategy))
                 .limit(heightOfLadder.height())
                 .collect(Collectors.toList());
@@ -20,6 +21,10 @@ public class Ladder {
 
     public static Ladder newLadder(HeightOfLadder heightOfLadder, PlayersAndWinningContents playersAndWinningContents, LinkStrategy linkStrategy) {
         return new Ladder(heightOfLadder, playersAndWinningContents, linkStrategy);
+    }
+
+    public PlayersAndWinningContents playersAndWinningContents() {
+        return playersAndWinningContents;
     }
 
     public List<Line> lines() {
