+++ src/main/java/laddergame/domain/ladder/Line.java
@@ -1,6 +1,5 @@
 package laddergame.domain.ladder;
 
-import laddergame.domain.PlayersAndWinningContents;
 import laddergame.domain.ladder.strategy.LinkStrategy;
 
 import java.util.*;
