+++ src/main/java/laddergame/domain/ladder/WinningContent.java
@@ -1,8 +1,8 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 import java.util.Objects;
 
-import static laddergame.domain.PlayersAndWinningContents.MAX_LENGTH_OF_PLAYER_AND_WINNING_CONTENT;
+import static laddergame.domain.ladder.PlayersAndWinningContents.MAX_LENGTH_OF_PLAYER_AND_WINNING_CONTENT;
 import static laddergame.exception.ExceptionMessage.WRONG_WINNING_MESSAGE;
 import static laddergame.util.Validator.isNonBlank;
