+++ src/main/java/laddergame/domain/ladder/WinningContents.java
@@ -1,11 +1,11 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static laddergame.domain.PlayersAndWinningContents.MIN_NUMBER_OF_PLAYERS_AND_WINNING_CONTENTS;
+import static laddergame.domain.ladder.PlayersAndWinningContents.MIN_NUMBER_OF_PLAYERS_AND_WINNING_CONTENTS;
 import static laddergame.exception.ExceptionMessage.WRONG_SIZE_OF_WINNINGS_MESSAGE;
 
 public class WinningContents {
