+++ src/test/java/laddergame/domain/ladder/LineTest.java
@@ -1,8 +1,5 @@
 package laddergame.domain.ladder;
 
-import laddergame.domain.Players;
-import laddergame.domain.PlayersAndWinningContents;
-import laddergame.domain.WinningContents;
 import laddergame.domain.ladder.strategy.LinkStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
